Add student course selection system with data initialization
Major Changes:
- Add complete student-course system (backend + frontend)
- Migrate from JPA to MyBatis Plus for both systems
- Add comprehensive test data initialization scripts
- Add batch scripts for easy startup

Student Course System:
- Backend: Spring Boot 3.2 + MyBatis Plus 3.5.5
- Frontend: Vue 3 + TypeScript + Element Plus
- Features: Student login, course selection, enrollment management
- JWT authentication for secure access
- Age group based course filtering

Database & Data:
- Add init-student-data.sql with 160+ course schedules
- 28 time slots (Mon-Sun, 4 slots/day)
- 8 classrooms with facilities
- 16 courses across 4 age groups
- 18 test students with credentials
- Add import-init-data.bat for easy data import

Documentation:
- Add 快速开始.md - Quick start guide
- Add 数据初始化说明.md - Detailed data setup instructions
- Add MYBATIS_PLUS_MIGRATION.md - JPA to MyBatis Plus migration guide
- Update CLAUDE.md with complete project documentation

Teacher Scheduling System Updates:
- Migrate from JPA repositories to MyBatis Plus mappers
- Add MyBatis Plus configuration and auto-fill handlers
- Update all service implementations for MyBatis Plus
- Add frontend components for schedule management
- Add data conflict query view

Project Structure:
- Add root pom.xml for multi-module management
- Add start-all.bat, start-backends.bat, start-frontends.bat
- Clean up deprecated teacher-f1, teacher-f2, old student-course dirs
- Organize under student-course/ directory with proper structure

This commit resolves the issue where students could not see courses after login
due to missing test data in the database.

+++ teacher-scheduling-system/backend/src/main/java/com/yl/paike/teacher/dto/PageResult.java
@@ -12,7 +12,8 @@ import java.util.List;
 public class PageResult<T> {
     private List<T> content;
     private Long totalElements;
-    private Integer totalPages;
+    // 改为 long，兼容不同处传 long/int
+    private long totalPages;
     private Integer page;
     private Integer size;
 }
